+++ api/v1alpha1/podnetworkchaos_types.go
@@ -188,7 +188,7 @@ type TcParameter struct {
 	Bandwidth *BandwidthSpec `json:"bandwidth,omitempty"`
 
 	// Rate represents the detail about rate control action
-	// +ui:form:when=action=='rate'
+	// +ui:form:ignore
 	// +optional
 	Rate *RateSpec `json:"rate,omitempty"`
 }