+++ ui/app/src/IntlProvider.tsx
@@ -14,16 +14,16 @@
  * limitations under the License.
  *
  */
-import flat from 'flat'
-import messages from 'i18n/messages'
+import messages from '@/i18n/messages'
+import { flatten } from 'flat'
 import { useMemo } from 'react'
 import { IntlProvider as ReactIntlProvider } from 'react-intl'
 
-import { useStoreSelector } from 'store'
+import { useSystemStore } from './zustand/system'
 
-const IntlProvider: React.FC = ({ children }) => {
-  const { lang } = useStoreSelector((state) => state.settings)
-  const intlMessages = useMemo<Record<string, string>>(() => flat(messages[lang]), [lang])
+const IntlProvider: ReactFCWithChildren = ({ children }) => {
+  const lang = useSystemStore((state) => state.lang)
+  const intlMessages = useMemo<Record<string, string>>(() => flatten(messages[lang]), [lang])
 
   return (
     <ReactIntlProvider messages={intlMessages} locale={lang} defaultLocale="en">