+++ ui/app/src/components/FormField/TextTextField.tsx
@@ -14,13 +14,12 @@
  * limitations under the License.
  *
  */
+import Space from '@/mui-extends/Space'
 import AddCircleTwoToneIcon from '@mui/icons-material/AddCircleTwoTone'
 import RemoveCircleTwoToneIcon from '@mui/icons-material/RemoveCircleTwoTone'
 import { Box, Button, FormHelperText, IconButton, Typography } from '@mui/material'
 import { getIn, useFormikContext } from 'formik'
 import _ from 'lodash'
-
-import Space from '@ui/mui-extends/esm/Space'
 
 import LabelField from './LabelField'
 import TextField from './TextField'