+++ ui/app/src/components/NewExperimentNext/Step3.tsx
@@ -14,39 +14,37 @@
  * limitations under the License.
  *
  */
+import Paper from '@/mui-extends/Paper'
+import PaperTop from '@/mui-extends/PaperTop'
+import { usePostExperiments } from '@/openapi'
+import { useComponentActions } from '@/zustand/component'
+import { useExperimentActions, useExperimentStore } from '@/zustand/experiment'
+import { useSettingStore } from '@/zustand/setting'
 import DoneAllIcon from '@mui/icons-material/DoneAll'
 import { Box, Typography } from '@mui/material'
-import { usePostExperiments } from 'openapi'
 import { useIntl } from 'react-intl'
-import { useNavigate } from 'react-router-dom'
+import { useNavigate } from 'react-router'
 
-import Paper from '@ui/mui-extends/esm/Paper'
-import PaperTop from '@ui/mui-extends/esm/PaperTop'
+import { Submit } from '@/components/FormField'
+import { type ExperimentKind } from '@/components/NewExperiment/types'
+import i18n from '@/components/T'
 
-import { useStoreDispatch, useStoreSelector } from 'store'
-
-import { resetNewExperiment } from 'slices/experiments'
-import { setAlert } from 'slices/globalStatus'
-
-import { Submit } from 'components/FormField'
-import { ExperimentKind } from 'components/NewExperiment/types'
-import i18n from 'components/T'
-
-import { parseSubmit } from 'lib/formikhelpers'
+import { parseSubmit } from '@/lib/formikhelpers'
 
 interface Step3Props {
   onSubmit?: (parsedValues: any) => void
   inSchedule?: boolean
 }
 
-const Step3: React.FC<Step3Props> = ({ onSubmit, inSchedule }) => {
+const Step3: ReactFCWithChildren<Step3Props> = ({ onSubmit, inSchedule }) => {
   const navigate = useNavigate()
   const intl = useIntl()
 
-  const state = useStoreSelector((state) => state)
-  const { step1, step2, kindAction, env, basic, spec } = state.experiments
-  const { debugMode, useNewPhysicalMachine } = state.settings
-  const dispatch = useStoreDispatch()
+  const { setAlert } = useComponentActions()
+
+  const { step1, step2, kindAction, env, basic, spec } = useExperimentStore()
+  const { debugMode, useNewPhysicalMachine } = useSettingStore()
+  const reset = useExperimentActions().reset
 
   const { mutateAsync } = usePostExperiments()
 
@@ -61,11 +59,11 @@ const Step3: React.FC<Step3Props> = ({ onSubmit, inSchedule }) => {
           ...spec,
         },
       },
-      { inSchedule, useNewPhysicalMachine }
+      { inSchedule, useNewPhysicalMachine },
     )
 
-    if (process.env.NODE_ENV === 'development' || debugMode) {
-      console.debug('Here is the experiment you are going to submit:', JSON.stringify(parsedValues, null, 2))
+    if (import.meta.env.DEV || debugMode) {
+      console.info('Here is the experiment you are going to submit:', JSON.stringify(parsedValues, null, 2))
     }
 
     if (!debugMode) {
@@ -76,28 +74,24 @@ const Step3: React.FC<Step3Props> = ({ onSubmit, inSchedule }) => {
           data: parsedValues,
         })
           .then(() => {
-            dispatch(
-              setAlert({
-                type: 'success',
-                message: i18n('confirm.success.create', intl),
-              })
-            )
+            setAlert({
+              type: 'success',
+              message: i18n('confirm.success.create', intl),
+            })
 
-            dispatch(resetNewExperiment())
+            reset()
 
             navigate('/experiments')
           })
           .catch((error) => {
-            if (process.env.NODE_ENV === 'development') {
+            if (import.meta.env.DEV) {
               console.error('Error submitting experiment:', error.response)
             }
 
-            dispatch(
-              setAlert({
-                type: 'error',
-                message: error.response.data.message,
-              })
-            )
+            setAlert({
+              type: 'error',
+              message: error.response.data.message,
+            })
           })
       }
     }