+++ ui/app/src/components/YAML/index.tsx
@@ -14,25 +14,22 @@
  * limitations under the License.
  *
  */
+import { useComponentActions } from '@/zustand/component'
 import FileOpenIcon from '@mui/icons-material/FileOpen'
 import LoadingButton, { LoadingButtonProps } from '@mui/lab/LoadingButton'
 import { useState } from 'react'
 
-import { useStoreDispatch } from 'store'
-
-import { setAlert } from 'slices/globalStatus'
-
-import { T } from 'components/T'
+import { T } from '@/components/T'
 
 interface YAMLProps {
   callback: (y: any) => void
   ButtonProps?: LoadingButtonProps<'label'>
 }
 
-const YAML: React.FC<YAMLProps> = ({ children, callback, ButtonProps }) => {
+const YAML: ReactFCWithChildren<YAMLProps> = ({ children, callback, ButtonProps }) => {
   const [loading, setLoading] = useState(false)
 
-  const dispatch = useStoreDispatch()
+  const { setAlert } = useComponentActions()
 
   const handleUploadYAML = (e: React.ChangeEvent<HTMLInputElement>) => {
     setLoading(true)
@@ -45,12 +42,10 @@ const YAML: React.FC<YAMLProps> = ({ children, callback, ButtonProps }) => {
 
       callback(y)
 
-      dispatch(
-        setAlert({
-          type: 'success',
-          message: <T id="confirm.success.load" />,
-        })
-      )
+      setAlert({
+        type: 'success',
+        message: <T id="confirm.success.load" />,
+      })
 
       setLoading(false)
     }